update data types in statistic, add user languages

+++ api/server/src/models/tagstatistic.js
@@ -2,10 +2,10 @@ module.exports = (sequelize, DataTypes) => {
   const TagStatistic = sequelize.define('TagStatistic', {
     TagId: DataTypes.INTEGER,
     name: DataTypes.STRING,
-    totalAnswers: DataTypes.INTEGER,
-    correctAnswers: DataTypes.INTEGER,
-    averageTimeOfCorrectAnswers: DataTypes.INTEGER,
-    averageTimeOfIncorrectAnswers: DataTypes.INTEGER,
+    totalAnswers: DataTypes.FLOAT,
+    correctAnswers: DataTypes.FLOAT,
+    averageTimeOfCorrectAnswers: DataTypes.FLOAT,
+    averageTimeOfIncorrectAnswers: DataTypes.FLOAT,
     percentile95OfCorrect: DataTypes.FLOAT,
     percentile95OfIncorrect: DataTypes.FLOAT,
   }, {});
